Finished tools to import and export custom draft classes.
I noticed a bug when I try to scroll through players.  If I open up a franchise file, then check the "Draft Class" and some position filter buttons in the lower left corner of the playerEditControl, then start scrolling through, eventually it will get to some player, throw an exception, and put me back at the start of the list.  If I scroll through backward, it seems to error on the same player again.  Maybe it's getting to the end of the list of players and returning NULL with your custom iteration functions?

Anyway, take a look and see if you can fix this bug, test it a bit, then let's rebuild and release the next version, which I guess will be Beta 4.