#45 Removed call to TDBCompactDatabase except for when shutting down the EditorModel. This should prevent the renumbering of the record numbers which is causing massive corruptions.
#32
Added an additional collection to store any records that we have marked as deleted in the program. If we unset the deleted flag of a record that exists in this new deletedRecordList then it is added back into the recordList.

All the methods that use GetRecords() should now not have to check for deleted flags as they should be in that list. I havent changed these yet as it shouldn't matter as there would be no records that are marked as deleted.


Need to test thoroughly now.!!