Training Camp BETA complete!!
Would like to run testing on setup.exe prior to release.

Colin, as we discussed the tune.txt from the trunk needs to be installed into 
installDirectory + "\\Conditioning\\TrainingCamp\\" 

I'm typing up a doc I plan on pasting into the FF.com post when the file gets uploaded. If you want me to up load, I'll be sure to change the authors this time. I didn't know that was an option the last upload I made.